+++ CHANGELOG.md
@@ -9,6 +9,7 @@ PdfOutlineItem, PdfOutlines, PdfAnnotationActionBase, PdfAnnotationLink PdfAnnot
 - PdfDocument: Added GetFieldsIterator()
 - PdfPage: Added GetFieldsIterator()
 - PdfSignature: Added TryGetPreviousRevision()
+- PdfCanvas: Added CopyContentsTo()
 
 ## Version 0.10.3
 - Fixed big performance regression introduced in 0.10, see #108